refactor: extract BraceClassifier from parse_one_statement
Move four brace-detection helpers into a dedicated module with a single
classify() entry point. Zero behavior change.
